feat: streamline navigation and add emerald accents
